LLM/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.28.0
python-docx>=0.8.11
PyPDF2>=3.0.0
python-pptx>=0.6.21
moviepy==2.2.1
Pillow>=9.0.0
gTTS>=2.3.0
pygame>=2.1.0
numpy>=1.21.0